// src/lib/foxtrickPsico.d.ts
export const foxtrickPsico: {
  getPrediction: (...args: unknown[]) => {
    maxSkill?: unknown;
    limit?: unknown;
    isGK?: unknown;
    undef?: unknown;
    formHigh?: unknown;
    formAvg?: unknown;
    formLow?: unknown;
    wageHigh?: unknown;
    wageAvg?: unknown;
    wageLow?: unknown;
  } | null;
};